% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_effect_quintiles.R
\name{estimate_effect_quintiles}
\alias{estimate_effect_quintiles}
\title{Estimating the propensity score model based on subclassification for replication Brookhart et al. (2006)}
\usage{
estimate_effect_quintiles(PS, data)
}
\arguments{
\item{PS}{The propensity score, as estimated by the function estimate_ps().}

\item{data}{The dataset containing the exposure variable A, outcome variable Y and relevant covariates.}
}
\value{
Returns the exposure effect, estimated within strata defined by quintiles of the propensity score and then averaged across strata. Note that NAs are removed in computing the mean;
NAs occur in subsets of the data that include non-exposed or exposed only, i.e., in separated datasets. Stores potential warnings for all models fitted in data subsets.
}
\description{
Estimating the propensity score model based on subclassification for replication Brookhart et al. (2006)
}
