% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ps.R
\name{estimate_ps}
\alias{estimate_ps}
\title{Estimating the propensity score model for replication Brookhart et al. (2006)}
\usage{
estimate_ps(ps_model_covs, data)
}
\arguments{
\item{ps_model_covs}{Character vector of covariates that should be included in the propensity score model.}

\item{data}{The dataset containing the exposure variable A, outcome variable Y and covariates.}
}
\value{
Provides a propensity score and c-statistic of the propensity score model. Stores warnings that occurred during propensity model fitting.
}
\description{
Estimating the propensity score model for replication Brookhart et al. (2006)
}
