% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{gen_data}
\alias{gen_data}
\title{Generate simulation data for replication Brookhart et al. (2006)}
\usage{
gen_data(
  seed,
  nobs,
  beta0,
  beta1,
  beta2,
  beta3,
  alpha0,
  alpha1,
  alpha2,
  alpha3,
  alpha4 = 0.5,
  trans = identity
)
}
\arguments{
\item{seed}{The seed used to generate a dataset.}

\item{nobs}{The number of observations of the generated dataset.}

\item{beta0}{The intercept of the model for the conditional mean of exposure A on a log-scale.}

\item{beta1}{The conditional association between covariate X1 and exposure A on a log-scale.}

\item{beta2}{The conditional association between covariate X2 and exposure A on a log-scale.}

\item{beta3}{The conditional association between covariate X3 and exposure A on a log-scale.}

\item{alpha0}{The intercept of the model for the conditional mean of outcome Y on a log-scale.}

\item{alpha1}{The conditional association between covariate X1 and outcome Y on a log-scale.}

\item{alpha2}{The conditional association between covariate X2 and outcome Y on a log-scale.}

\item{alpha3}{The conditional association between covariate X3 and outcome Y on a log-scale.}

\item{alpha4}{The conditional association between exposure A and outcome Y on a log-scale. Set to 0.5 for all scenarios in this replication.}

\item{trans}{Transformations of covariate X1 in generating the outcome, Y. For experiment 1, use trans = identity; for experiment 2, use trans = function(x) 1/(1+exp(-3 * x)) * -.5
on a log-scale.}
}
\value{
The function returns a simulated dataset containing a dichotomous exposure, an outcome with a Poisson distribution, and continuous independent confounders. Both simulation experiment 1 and experiment 2 of the manuscript employ the same data-generating process.
}
\description{
Generate simulation data for replication Brookhart et al. (2006)
}
